// libs/supabase/schema.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export interface Database {
// ...
  public: {
    Tables: {
      gists: {
        Row: {
          content: string
          created_at: string
          description: string
          id: string
          is_paid: boolean
          lang: string
          price: number
          profile_uid: string
          title: string
        }
        Insert: {
          content: string
          created_at?: string
          description: string
          id: string
          is_paid: boolean
          lang: string
          price: number
          profile_uid: string
          title: string
        }
        Update: {
          content?: string
          created_at?: string
          description?: string
          id?: string
          is_paid?: boolean
          lang?: string
          price?: number
          profile_uid?: string
          title?: string
        }
        Relationships: [
          {
            foreignKeyName: 'gists_profile_uid_fkey'
            columns: ['profile_uid']
            isOneToOne: false
            referencedRelation: 'profiles'
            referencedColumns: ['id']
          },
        ]
      }
      profiles: {
        Row: {
          address: Json | null
          avatar_url: string
          bio: string | null
          created_at: string
          email: string
          id: string
          name: string
          payment_connected_account: string | null
          phone: string | null
          site: string | null
          username: string
        }
        Insert: {
          address?: Json | null
          avatar_url: string
          bio?: string | null
          created_at?: string
          email: string
          id: string
          name: string
          payment_connected_account?: string | null
          phone?: string | null
          site?: string | null
          username: string
        }
        Update: {
          address?: Json | null
          avatar_url?: string
          bio?: string | null
          created_at?: string
          email?: string
          id?: string
          name?: string
          payment_connected_account?: string | null
          phone?: string | null
          site?: string | null
          username?: string
        }
        Relationships: [
          {
            foreignKeyName: 'profiles_id_fkey'
            columns: ['id']
            isOneToOne: true
            referencedRelation: 'users'
            referencedColumns: ['id']
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never;
    }
    Functions: {
      [_ in never]: never;
    }
    Enums: {
      [_ in never]: never;
    }
    CompositeTypes: {
      [_ in never]: never;
    }
  }
// ...
}